requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-texttospeech
pydantic
google-cloud-firestore
python-dotenv
